+++ src/backend/joanie/urls.py
@@ -13,6 +13,8 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
+from django.conf import settings
+from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
 
@@ -25,9 +27,17 @@ router.register("courses", api.CourseViewSet, basename="courses")
 router.register("enrollments", api.EnrollmentViewSet, basename="enrollments")
 router.register("orders", api.OrderViewSet, basename="orders")
 
-urlpatterns = [
-    path("admin/", admin.site.urls),
-    path("api/addresses/", api.AddressView.as_view()),
-    path("api/addresses/<str:address_uid>/", api.AddressView.as_view()),
-    path("api/", include(router.urls)),
-]
+urlpatterns = (
+    [
+        path("admin/", admin.site.urls),
+        path("api/addresses/", api.AddressView.as_view()),
+        path("api/addresses/<str:address_uid>/", api.AddressView.as_view()),
+        path("api/", include(router.urls)),
+        path("api/documents/", include("marion.urls")),
+    ]
+    + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+    + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
+)
+
+if settings.DEBUG:
+    urlpatterns += [path("__debug__/", include("marion.urls.debug"))]